refactor(dye_liquifier): change colors amount to multiplier

+++ src/main/java/wootrevived/woot/recipes/dye_liquifier/DyeLiquifierRecipeBuilder.java
@@ -10,11 +10,11 @@ import java.util.List;
 
 public class DyeLiquifierRecipeBuilder {
     private int energy;
-    private int red;
-    private int yellow;
-    private int blue;
-    private int white;
-    private int multiply;
+    private float red;
+    private float yellow;
+    private float blue;
+    private float white;
+    private float multiply;
     private Ingredient ingredient;
 
     protected DyeLiquifierRecipeBuilder() {
@@ -30,27 +30,27 @@ public class DyeLiquifierRecipeBuilder {
         return this;
     }
 
-    public DyeLiquifierRecipeBuilder red(int red){
+    public DyeLiquifierRecipeBuilder red(float red){
         this.red = red;
         return this;
     }
 
-    public DyeLiquifierRecipeBuilder yellow(int yellow){
+    public DyeLiquifierRecipeBuilder yellow(float yellow){
         this.yellow = yellow;
         return this;
     }
 
-    public DyeLiquifierRecipeBuilder blue(int blue){
+    public DyeLiquifierRecipeBuilder blue(float blue){
         this.blue = blue;
         return this;
     }
 
-    public DyeLiquifierRecipeBuilder white(int white){
+    public DyeLiquifierRecipeBuilder white(float white){
         this.white = white;
         return this;
     }
 
-    public DyeLiquifierRecipeBuilder multiply(int multiply){
+    public DyeLiquifierRecipeBuilder multiply(float multiply){
         this.multiply = multiply;
         return this;
     }
